+ import albums from media storage

+++ data/src/main/java/com/github/anrimian/musicplayer/data/storage/providers/albums/StorageAlbumsProvider.java
@@ -62,8 +62,8 @@ public class StorageAlbumsProvider {
                 cursorWrapper.getString(Albums.ALBUM_KEY),
                 cursorWrapper.getString(Albums.ARTIST),
                 cursorWrapper.getLong(Albums.ARTIST_ID),
-                cursorWrapper.getLong(Albums.FIRST_YEAR),
-                cursorWrapper.getLong(Albums.LAST_YEAR)
+                cursorWrapper.getInt(Albums.FIRST_YEAR),
+                cursorWrapper.getInt(Albums.LAST_YEAR)
         );
     }
 }
